some tweaks

+++ deploy-dotvim.bash
@@ -2,6 +2,14 @@
 
 cd
 #git clone https://github.com/elts/dotvim.git ~/.vim
-git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
-ln -s .vim/.vimrc .vimrc
+if [ ! -d ~/.vim/bundle/vundle ]
+then
+    echo "Cloning Vundle from github"
+    git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
+fi
+if [ ! -f ~/.vimrc ]
+then
+    echo "Making a symlink to .vimrc"
+    ln -s .vim/.vimrc .vimrc
+fi
 vim +BundleInstall +qall
